niks

+++ src/components/admin-panel/book-overview/BookRow.vue
@@ -13,14 +13,15 @@
       </button>
     </router-link>
 
-    <button class="float-right text-white bg-blue-500 px-4 py-2 m-2 h-fit rounded-md transition ease-in-out delay-50 hover:-translate-y-1 hover:scale-110 hover:bg-indigo-500 duration-300" 
-      @click="deleteBook(book.id)">Delete</button>
-
     <router-link :to="{ name: 'book-detail', params: { id: book.id } }" class="hover:text-gray-200">
       <button
         class="float-right text-white bg-blue-500 px-4 py-2 m-2 h-fit rounded-md transition ease-in-out delay-50 hover:-translate-y-1 hover:scale-110 hover:bg-indigo-500 duration-300">Bekijk
       </button>
     </router-link>
+    
+    <button v-on:click="Archive()"
+        class="float-right text-white bg-blue-500 px-4 py-2 m-2 h-fit rounded-md transition ease-in-out delay-50 hover:-translate-y-1 hover:scale-110 hover:bg-indigo-500 duration-300">Archiveer
+    </button>
   </div>
 
 
@@ -37,18 +38,37 @@ export default {
 
     methods: {
       deleteBook(id) {
-
-        axios.delete(`http://localhost:8080/book/delete/${id}`)
+        if(window.confirm("Weet je zeker dat je dit boek wilt verwijderen?")){
+          axios.delete(`http://localhost:8080/book/delete/${id}`)
           .then(response => {
-            console.log('Book deleted:', response.data);
-            alert('Book has been deleted')
+            console.log('Verwijderde boek:', response.data);
+            alert('Het boek is verwijderd')
             window.location.reload()
-            // Remove the deleted book from the books array
           })
           .catch(error => {
             console.log(error);
           })
+        }
+        else{
+          alert("Het proces is geannuleerd")
+        }
+        
       },
+      Archive(){
+      if (confirm("Weet je zeker dat je deze boek wilt archiveren? Het boek kan dan niet meer uitgeleend worden.")){
+        axios.put("http://localhost:8080/book/archive/" + this.book.id)
+      .then(response => {
+        console.log(response.data)
+        if (response.data == true){
+          alert("Boek " + this.book.title + " is gearchiveerd")
+          window.location.reload();
+        }
+      })
+      .catch(error => {
+        console.log(error);
+          })
+        }
+      }
     },
 }
 </script>
